proste zbieranie surowców, do przerobienia, najelpiej podzielic maper na kwadraty

+++ src/hud/window/top/TopPanel.cpp
@@ -52,7 +52,7 @@ void TopPanel::update(Resources* resources) {
 	if (resources->hasChanged()) {
 
 		short size = resources->getSize();
-		float* values = resources->getValues();
+		double* values = resources->getValues();
 		for (int i = 0; i < size; ++i) {
 			elements[i]->setText(String((int)values[i]));
 		}
